+++ apps/server/agents/base_agent.py
@@ -2,7 +2,7 @@ from uuid import UUID
 from typings.user import UserOutput
 from typings.account import AccountOutput
 
-class L3Base:
+class BaseAgent:
     def __init__(self,  user: UserOutput, account: AccountOutput, session_id: UUID):
         self.user = user
         self.account = account
